+++ src/pygpsclient/banner_frame.py
@@ -405,11 +405,11 @@ class BannerFrame(Frame):
         else:
             self._time.set(tim)
 
-    def _update_pos(self, deg_format, units):
+    def _update_pos(self, pos_format, units):
         """
-        Update position
+        Update position.
 
-        :param str disp_format: position display format as string (DMS, DMM, DDD, ECEF)
+        :param str pos_format: position display format as string (DMS, DMM, DDD, ECEF)
         :param str units: distance units as string (UMM, UMK, UI, UIK)
         """
 
@@ -419,34 +419,35 @@ class BannerFrame(Frame):
         self._lbl_llat.config(text="lat:")
         self._lbl_llon.config(text="lon:")
         self._lbl_lalt.config(text="alt:")
-        alt_u = "m"
+        alt_u = "ft" if units in (UI, UIK) else "m"
 
-        if isinstance(lat, (int, float)) and isinstance(lon, (int, float)):
-            if deg_format == ECEF and isinstance(alt, (int, float)):
+        try:
+            if pos_format == ECEF:
                 lat, lon, alt = llh2ecef(lat, lon, alt)
                 if units in (UI, UIK):
                     lat, lon = (m2ft(x) for x in (lat, lon))
                 self._lbl_llat.config(text="X:")
                 self._lbl_llon.config(text="Y:")
                 self._lbl_lalt.config(text="Z:")
-            elif deg_format == DMS:
-                lat, lon = latlon2dms(lat, lon)
-            elif deg_format == DMM:
-                lat, lon = latlon2dmm(lat, lon)
-            self._lat.set(f"{lat:<15}")
-            self._lon.set(f"{lon:<15}")
-        else:
+                self._lat.set(f"{lat:.4f}")
+                self._lon.set(f"{lon:.4f}")
+            else:
+                deg_f = "<15"
+                if pos_format == DMS:
+                    lat, lon = latlon2dms(lat, lon)
+                elif pos_format == DMM:
+                    lat, lon = latlon2dmm(lat, lon)
+                else:
+                    deg_f = ".8f"
+                if units in (UI, UIK):
+                    alt = m2ft(alt)
+                self._lat.set(f"{lat:{deg_f}}")
+                self._lon.set(f"{lon:{deg_f}}")
+            self._alt.set(f"{alt:.4f}")
+            self._alt_u.set(f"{alt_u:<2}")
+        except (TypeError, ValueError):
             self._lat.set("N/A            ")
             self._lon.set("N/A            ")
-
-        if isinstance(alt, (int, float)):
-            if units in (UI, UIK):
-                alt = m2ft(alt)
-                alt_u = "ft"
-            fmt = "<15" if deg_format == ECEF else ".3f"
-            self._alt.set(f"{alt:{fmt}}")
-            self._alt_u.set(f"{alt_u:<2}")
-        else:
             self._alt.set("N/A  ")
             self._alt_u.set("  ")